==> 00_scraping.ipynb <==
from scraping import CianScraper
from scraping import By

import pandas as pd

scraper = CianScraper()

scraper.start('https://kaliningrad.cian.ru/kupit-kvartiru/')

data = scraper.collect_data()

10636

df = pd.DataFrame(data)

df['link'].nunique()

df.to_csv('output/cian_data.csv', index=False)
