==> 01_preprocessing.ipynb <==
import datetime

import pandas as pd
import numpy as np
from functools import reduce

import preprocessing as prep

df = pd.read_csv('output/cian_data.csv')

first, second = df.iloc[:10636], df.iloc[10636:]

first['dt'] = first['dt'].map(lambda x: prep.get_dt(x, datetime.date(2022, 11, 26)))
second['dt'] = second['dt'].map(lambda x: prep.get_dt(x, datetime.date(2022, 11, 27)))

df = pd.concat([first, second])

df = prep.first_preprocessing(df)

funcs = [
    prep.get_nrooms_type,
    prep.get_price_area,
    prep.get_address_components,
    prep.get_cian_description,
    prep.get_availability,
    prep.get_floors,
]
df = reduce(lambda x, y: x.join(x.apply(y, axis=1)), [df, *funcs])
prep.rename_drop_columns(df)

pd.set_option('display.max_columns', None)

df

df.to_csv('output/parsed_data.csv', index=False)
